Remove Token class
